#70 refac: daily 삭제, ContentView 삭제, RoundedShape 파일분리
